Modules/pyproject.toml
[tool.poetry]
name = "modules"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "*"
configparser = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
